Split code into specific modules for clarity
