+++ apps/transport/lib/transport/history/fetcher_mock.ex
@@ -1,0 +1,10 @@
+defmodule Transport.History.Fetcher.Mock do
+  @behaviour Transport.History.Fetcher
+
+  @moduledoc """
+  A default implementation returning an empty history, useful as a default
+  implementation for tests unrelated to history.
+  """
+
+  def history_resources(%DB.Dataset{}), do: []
+end
